requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.27.0
google-generativeai>=0.8.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.24.0

# Agents (teacher / verifier / debate graph)
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langchain-openai>=0.2.0
litellm>=1.63.0
langgraph>=0.2.0
langchain-community>=0.3.0
langchain-experimental>=0.3.0
playwright>=1.40.0
typing_extensions>=4.0.0
